HomeController の名称を変更し、トップページ用のコントローラーとした
